requirements.txt
Django==4.0.4
mysql-connector==2.2.9
virtualenv==20.0.31
dj-database-url==0.5.0
gunicorn==20.0.4
psycopg2==2.9.3
whitenoise==5.2.0
Pillow==9.1.0
requests==2.24.0

django-modeltranslation~=0.18.2